Get tests working for icesat2_boreal pipeline
