+++ image_builder.py
@@ -52,6 +52,7 @@ def build_and_push(dockerfile: str, tag: str, pipeline_id: str):
     push_log = client.images.push(ecr_repo_name, tag="latest")
     logging.debug(push_log)
 
+
 pipeline_path = f"./stactools_pipelines/pipelines/{pipeline_name}"
 with open(f"{pipeline_path}/config.yaml") as f:
     config = yaml.safe_load(f)
@@ -60,14 +61,20 @@ with open(f"{pipeline_path}/config.yaml") as f:
     client = docker.from_env()
     if pipeline.compute == "awslambda":
         if os.path.exists(f"{pipeline_path}/lambda.collection.Dockerfile"):
-            build_and_push(f"{pipeline_path}/lambda.collection.Dockerfile", pipeline.id, pipeline.id)
+            build_and_push(
+                f"{pipeline_path}/lambda.collection.Dockerfile",
+                pipeline.id,
+                pipeline.id,
+            )
         else:
             dockerfile = "./lambda.collection.Dockerfile"
             tag = f"{pipeline.id}-collection"
             build_and_push(dockerfile, tag, pipeline.id)
 
         if os.path.exists(f"{pipeline_path}/lambda.Dockerfile"):
-            build_and_push(f"{pipeline_path}/lambda.Dockerfile", pipeline.id, pipeline.id)
+            build_and_push(
+                f"{pipeline_path}/lambda.Dockerfile", pipeline.id, pipeline.id
+            )
         else:
             dockerfile = "./lambda.Dockerfile"
             tag = pipeline.id
