--- stactools_pipelines/pipelines/icesat2-boreal/test_collection.py
@@ -1,19 +1,0 @@
-import json
-
-import pytest
-
-from stactools_pipelines.pipelines import conftest
-from stactools_pipelines.pipelines.icesat2_boreal.collection import handler
-
-
-@pytest.mark.parametrize("pipeline_id", ["icesat2_boreal"])
-@pytest.mark.parametrize("module", ["collection"])
-def test_handler(mock_env, get_token, create_collection, requests):
-    handler({}, {})
-    get_token.assert_called_once()
-    create_collection.assert_called_once_with()
-    requests.post.assert_called_once_with(
-        url=f"{conftest.ingestor_url}/collections",
-        data=json.dumps(conftest.collection),
-        headers={"Authorization": f"bearer {conftest.token}"},
-    )
