+++ stactools_pipelines/pipelines/icesat2_boreal/config.yaml
@@ -1,5 +1,5 @@
 ---
-  id: "icesat2-boreal"
+  id: "icesat2_boreal"
   compute: "awslambda"
   secret_arn: "arn:aws:secretsmanager:us-west-2:916098889494:secret:MAAP-STAC-auth-dev/MAAP-workflows-EsykqB" # cognito secret ARN
   ingestor_url: "https://stac-ingestor.dit.maap-project.org"
