+++ stactools_pipelines/pipelines/icesat2_boreal/test_historic.py
@@ -1,9 +1,10 @@
+from unittest.mock import patch
+
 import pytest
 
 import stactools_pipelines.pipelines.conftest as conftest
 from stactools_pipelines.pipelines.icesat2_boreal.historic import (
     handler,
-    inventory_data,
 )
 
 inventory_location = "inventory_location"
@@ -18,14 +19,14 @@ def mock_inventory_env(monkeypatch):
 @pytest.fixture()
 def inventory_data():
     with patch(
-        "stactools_pipelines.pipelines.panama_copc.historic.inventory_data",
+        "stactools_pipelines.pipelines.icesat2_boreal.historic.inventory_data",
         autospec=True,
     ) as m:
         m.return_value = [key]
         yield m
 
 
-@pytest.mark.parametrize("pipeline_id", ["panama-copc"])
+@pytest.mark.parametrize("pipeline_id", ["icesat2_boreal"])
 @pytest.mark.parametrize("query_value", [""])
 def test_handler(mock_env, mock_inventory_env, inventory_data, boto3):
     handler({}, {})
