+++ tox.ini
@@ -1,5 +1,6 @@
 [tox]
 envlist = py39
+isolated_build = True
 
 [testenv]
 extras = test
